🚚 move: 테스트코드 패키지 정리
