+++ src/test/java/com/project/cheerha/user/verification/CreatePasswordResetTokenTest.java
@@ -1,4 +1,4 @@
-package com.project.cheerha.user;
+package com.project.cheerha.user.verification;
 
 import com.project.cheerha.common.exception.data.NotFoundException;
 import com.project.cheerha.common.redis.email.EmailTokenService;
